# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/cart/items/{customerUserId}": {
      "get": {
        "operationId": "CartController_getCartItems",
        "parameters": [
          {
            "name": "customerUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "API接口",
        "tags": [
          "购物车管理"
        ]
      },
      "post": {
        "operationId": "CartController_addCartItem",
        "parameters": [
          {
            "name": "customerUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "创建资源的请求数据",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartItemResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "创建资源",
        "tags": [
          "购物车管理"
        ]
      },
      "delete": {
        "description": "根据商品ID列表批量删除用户购物车中的商品",
        "operationId": "CartController_removeCartItems",
        "parameters": [
          {
            "name": "customerUserId",
            "required": true,
            "in": "path",
            "description": "用户唯一标识符",
            "schema": {
              "example": "user_123456",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "要删除的购物车商品ID列表",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "批量删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "批量删除购物车商品",
        "tags": [
          "购物车管理"
        ]
      }
    },
    "/api/cart/selected/{customerUserId}": {
      "get": {
        "operationId": "CartController_getSelectedCartItems",
        "parameters": [
          {
            "name": "customerUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "API接口",
        "tags": [
          "购物车管理"
        ]
      },
      "delete": {
        "description": "删除用户购物车中所有已选中的商品",
        "operationId": "CartController_clearSelectedItems",
        "parameters": [
          {
            "name": "customerUserId",
            "required": true,
            "in": "path",
            "description": "用户唯一标识符",
            "schema": {
              "example": "user_123456",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "清空选中商品成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "清空选中商品",
        "tags": [
          "购物车管理"
        ]
      }
    },
    "/api/cart/items/{customerUserId}/{cartItemId}": {
      "put": {
        "description": "更新购物车中指定商品的数量、选中状态等信息",
        "operationId": "CartController_updateCartItem",
        "parameters": [
          {
            "name": "customerUserId",
            "required": true,
            "in": "path",
            "description": "用户唯一标识符",
            "schema": {
              "example": "user_123456",
              "type": "string"
            }
          },
          {
            "name": "cartItemId",
            "required": true,
            "in": "path",
            "description": "购物车商品唯一标识符",
            "schema": {
              "example": "cart_123456",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "更新商品的信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "购物车商品不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新购物车商品",
        "tags": [
          "购物车管理"
        ]
      },
      "delete": {
        "description": "从用户购物车中删除指定的商品",
        "operationId": "CartController_removeCartItem",
        "parameters": [
          {
            "name": "customerUserId",
            "required": true,
            "in": "path",
            "description": "用户唯一标识符",
            "schema": {
              "example": "user_123456",
              "type": "string"
            }
          },
          {
            "name": "cartItemId",
            "required": true,
            "in": "path",
            "description": "购物车商品唯一标识符",
            "schema": {
              "example": "cart_123456",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "购物车商品不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除购物车商品",
        "tags": [
          "购物车管理"
        ]
      }
    },
    "/api/cart/select-all/{customerUserId}": {
      "put": {
        "description": "批量设置用户购物车中所有商品的选中状态",
        "operationId": "CartController_updateAllSelectFlag",
        "parameters": [
          {
            "name": "customerUserId",
            "required": true,
            "in": "path",
            "description": "用户唯一标识符",
            "schema": {
              "example": "user_123456",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "选中状态标志",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "批量更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "批量更新选中状态",
        "tags": [
          "购物车管理"
        ]
      }
    },
    "/api/cart/summary/{customerUserId}": {
      "get": {
        "operationId": "CartController_getCartSummary",
        "parameters": [
          {
            "name": "customerUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "API接口",
        "tags": [
          "购物车管理"
        ]
      }
    },
    "/api/cart/count/{customerUserId}": {
      "get": {
        "operationId": "CartController_getCartItemCount",
        "parameters": [
          {
            "name": "customerUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "API接口",
        "tags": [
          "购物车管理"
        ]
      }
    }
  },
  "info": {
    "title": "Caddy Style Shopping API",
    "description": "\n        ## 电商系统 API 文档\n        \n        ### 功能模块\n        - **认证管理**: 用户注册、登录、JWT令牌管理\n        - **用户管理**: 用户信息、权限管理、地址管理\n        - **商品管理**: 商品CRUD、分类、库存管理\n        - **购物车管理**: 购物车操作、商品选择、批量操作\n        - **订单管理**: 订单创建、支付、状态跟踪\n        - **支付管理**: 多种支付方式、支付回调处理\n        - **系统管理**: 缓存管理、性能监控、健康检查\n        \n        ### 认证方式\n        使用 JWT Bearer Token 进行认证，请在请求头中添加：\n        `Authorization: Bearer <your-token>`\n        \n        ### 响应格式\n        所有API响应都遵循统一格式：\n        ```json\n        {\n          \"code\": 200,\n          \"message\": \"操作成功\",\n          \"data\": {},\n          \"timestamp\": \"2025-01-26T10:30:00Z\",\n          \"requestId\": \"req_123456789\"\n        }\n        ```\n        \n        ### 错误码说明\n        - **1xxx**: 通用错误（参数错误、资源不存在等）\n        - **2xxx**: 认证错误（未授权、令牌过期等）\n        - **3xxx**: 业务错误（库存不足、订单状态错误等）\n        - **5xxx**: 系统错误（服务器错误、数据库错误等）\n      ",
    "version": "1.0.0",
    "contact": {
      "name": "开发团队",
      "url": "https://github.com/your-org/caddy-style-shopping",
      "email": "dev@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发环境"
    },
    {
      "url": "https://api-staging.example.com",
      "description": "测试环境"
    },
    {
      "url": "https://api.example.com",
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "错误状态码",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "错误消息",
            "example": "请求参数错误"
          },
          "errors": {
            "description": "错误详情",
            "example": [
              "用户名不能为空",
              "密码长度至少6位"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timestamp": {
            "type": "string",
            "description": "时间戳",
            "example": "2025-01-26T10:30:00Z"
          },
          "path": {
            "type": "string",
            "description": "请求路径",
            "example": "/api/auth/login"
          },
          "requestId": {
            "type": "string",
            "description": "请求ID",
            "example": "req_123456789"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp"
        ]
      },
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "响应状态码",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "timestamp": {
            "type": "string",
            "description": "时间戳",
            "example": "2025-01-26T10:30:00Z"
          },
          "requestId": {
            "type": "string",
            "description": "请求ID",
            "example": "req_123456789"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp"
        ]
      },
...
      "CartItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "购物车商品ID",
            "example": "cart_1234567890_abc123"
          },
          "productId": {
            "type": "string",
            "description": "商品 SPU ID",
            "example": "prod_123456"
          },
          "productSkuId": {
            "type": "string",
            "description": "商品 SKU ID",
            "example": "sku_123456"
          },
          "productName": {
            "type": "string",
            "description": "商品名称",
            "example": "Reich 经典手袋"
          },
          "productBrand": {
            "type": "string",
            "description": "商品品牌",
            "example": "Reich"
          },
          "productPrice": {
            "type": "number",
            "description": "商品价格",
            "example": 2999.99
          },
          "productQuantity": {
            "type": "number",
            "description": "商品数量",
            "example": 1
          },
          "productPic": {
            "type": "string",
            "description": "商品图片URL",
            "example": "/images/products/bag-001.jpg"
          },
          "productAttribute": {
            "type": "string",
            "description": "商品规格属性",
            "example": "{\"color\":\"黑色\",\"size\":\"中号\"}"
          },
          "selectFlag": {
            "type": "boolean",
            "description": "是否选中",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "totalPrice": {
            "type": "number",
            "description": "商品总价",
            "example": 2999.99
          }
        },
        "required": [
          "id",
          "productId",
          "productSkuId",
          "productName",
          "productBrand",
          "productPrice",
          "productQuantity",
          "productPic",
          "productAttribute",
          "selectFlag",
          "createdAt",
          "updatedAt",
          "totalPrice"
        ]
      },
...
    }
  }
}